Retire recommended list view
